+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Fixed
+
+- Fix Slider drag stops on scroll now
+
 ## 3.15.6 (August 31, 2022)
 
 ### Changed